Move test-manager-base.ts to /query-testing/test-manager-base.ts
